Account for differences in API host
All tests were passing except this one, which after the refactor
is no longer expecting the right API host. Consider:

www.gov.uk/api/organisations/hm-revenue-customs
vs
http://www.dev.gov.uk/api/organisations/hm-revenue-customs

In practice it doesn't matter what the host is, it matters what
the path is. Elsewhere in the file we make a similar trade-off:
https://github.com/alphagov/gds-api-adapters/blob/9374ef667e9a30696617faf7c013bfdfecf9bce7/test/organisations_api_test.rb#L55